refactor: remove deprecated scripts/config, update all imports to src/config
- Update all scripts to import directly from src/config instead of scripts/config
- Update test files to import from src/config
- Remove deprecated scripts/config/addresses.ts
- Remove deprecated scripts/config/range.ts
- Remove empty scripts/config directory

All scripts now use the centralized configuration system directly,
eliminating the deprecated re-export layer.
