+++ scripts/execute-rebalance.ts
@@ -1,5 +1,5 @@
 import { ethers } from "hardhat";
-import { ARBITRUM_MAINNET } from "./config/addresses";
+import { ARBITRUM_MAINNET } from "../src/config/addresses";
 import { DeltaNeutralMonitor } from "../src/strategy/monitor";
 import { RebalanceManager } from "../src/strategy/rebalance";
 import { StrategyAction } from "../src/strategy/types";
