+++ scripts/read-orders.ts
@@ -1,6 +1,6 @@
 import { ethers } from "hardhat";
 import readerDeployment from "../deployments/Reader.arbitrum.json";
-import { ARBITRUM_MAINNET } from "./config/addresses";
+import { ARBITRUM_MAINNET } from "../src/config/addresses";
 
 async function main() {
   const [signer] = await ethers.getSigners();
